Bundle Framework7 and optimize CSS
Instead of borrowing from Framework7, this bundles it, but optimizes by
purging unused CSS. This is not only important to final bundle size, but
we don't want to style anything but the preview and not include body
styles, etc. as to not conflict with the app that is depending on us.