Restructure DOM and API
This introduces a new DOM structure to space actions out from the header
so it can be optional. A container is where the main content of the
Shortcut will live now.

This also changes the intializer API from functions to a class.

So now it will be intialized with this:

```
const preview = new ShortcutPreview({
	// options...
})
```