Implement Shortcut Details and Setup (#2)
This implements an info circle fixed at the top of the Shortcut preview that, when clicked, pops a modal that, for now, is custom.

This implements the Details and Setup tabs with segmented controls all functionality of which provided by Framework7, with some needing some dark mode holes patched up.